feat: add float rating and notes to books

+++ backend/reading/models.py
@@ -13,7 +13,7 @@ class Book(models.Model):
     author = models.CharField(max_length=255)
     status = models.CharField(max_length=10, choices=STATUS_CHOICES, default='want')
     owned = models.BooleanField(default=False)
-    rating = models.IntegerField(null=True, blank=True)
+    rating = models.FloatField(null=True, blank=True)
     notes = models.TextField(blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
